Improve DNS TXT client and server more
